added relationship for User and Book

+++ src/routers/book.js
@@ -1,9 +1,13 @@
 const express = require('express')
 const router = new express.Router()
 const Book = require('../models/book')
+const auth = require('../middleware/auth')
 
-router.post('/books', async (req, res) => {
-    const book = new Book(req.body)
+router.post('/books', auth, async (req, res) => {
+    const book = new Book({
+        ...req.body,
+        created_by: req.user._id
+    })
     
     try {
         await book.save()
@@ -13,21 +17,23 @@ router.post('/books', async (req, res) => {
     }
 })
 
-router.get('/books', async (req, res) => {
+router.get('/books', auth, async (req, res) => {
     
     try {
         const books = await Book.find({})
         res.status(200).send(books)
+
     } catch (e) {
         res.status(500).send(e)   
     }
 })
 
-router.get('/books/:id', async (req, res) => {
+router.get('/books/:id', auth, async (req, res) => {
     _id = req.params.id
 
     try {
-        const book = await Book.findById(_id)
+        const book = await Book.findOne({_id, created_by: req.user._id})
+
         if (!book) {
             res.status(404).send()
         }
@@ -37,7 +43,7 @@ router.get('/books/:id', async (req, res) => {
     }
 })
 
-router.patch('/books/:id', async (req, res) => {
+router.patch('/books/:id', auth, async (req, res) => {
     const updates = Object.keys(req.body)
     const allowedUpdates = ['name', 'description', 'rating']
     const isValidOperation = updates.every((update) => allowedUpdates.includes(update))
@@ -49,26 +55,25 @@ router.patch('/books/:id', async (req, res) => {
     _id = req.params.id
 
     try {
-        const book = await Book.findById(req.params.id)
-
-        updates.forEach((update) => book[update] = req.body[update])
-
-        await book.save()
+        const book = await Book.findOne({_id: req.params.id, created_by: req.user._id})
 
         if (!book) {
             return res.status(404).send()
         }
+
+        updates.forEach((update) => book[update] = req.body[update])
+        await book.save()
         res.status(200).send(book)
     } catch (e) {
         res.status(500).send(e)
     }
 })
 
-router.delete('/books/:id', async (req, res) => {
+router.delete('/books/:id', auth, async (req, res) => {
     _id = req.params.id
 
     try {
-        const book = await Book.findByIdAndDelete(_id)
+        const book = await Book.findByIdAndDelete({_id: req.params.id, created_by: req.user._id})
         if (!book) {
             res.status(404).send({error: 'Book not found'})
         }
